pyproject.toml
[project]
requires-python = "==3.14.*"
name = "bot"
version = "1.0.0"
description = ""
authors = [{ name = "BlindfoldedSurgery", email = "team@blindfolded.surgery" }]
classifiers = [
    "Private :: Do Not Upload",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "azure-core ==1.36.*",
    "azure-core-tracing-opentelemetry",
    "azure-cognitiveservices-speech ==1.46.*",
    "bs-config [dotenv] ==3.4.0",
    "bs-nats-updater ==3.0.0",
    "bs-state [redis] ==3.0.*",
    "click >=8, <9",
    "opentelemetry-api ==1.38.*",
    "opentelemetry-sdk ==1.38.*",
    "opentelemetry-exporter-otlp-proto-grpc ==1.38.*",
    "opentelemetry-instrumentation-asyncio",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-logging",
    "prep-rate-limiter [postgres,opentelemetry-postgres] ==8.1.2",
    "python-telegram-bot ==22.5",
    "sentry-sdk >=2, <3",
    "uvloop ==0.22.*",
]

[dependency-groups]
dev = [
    "mypy ==1.18.*",
    "pytest >=8, <9",
    "ruff ==0.14.3",
]

[[tool.uv.index]]
name = "internal"
url = "https://pypi.bjoernpetersen.net/simple"
explicit = true

[tool.uv.sources]
bs-config = { index = "internal" }
bs-nats-updater = { index = "internal" }
bs-state = { index = "internal" }
prep-rate-limiter = { index = "internal" }

[build-system]
requires = ["uv_build==0.9.*"]
build-backend = "uv_build"

[tool.mypy]
strict = true
warn_return_any = false

[[tool.mypy.overrides]]
module = "tests.*"
strict = false
check_untyped_defs = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "azure.*"
ignore_missing_imports = true

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "FURB",
    "G",
    "I001",
    "LOG",
    "PTH",
    "Q",
    "TC001",
    "TC002",
    "TC003",
    "TC004",
    "TC005",
    "UP",
]
ignore = [
    # redundant-open-modes
    "UP015",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = [
    "pydantic.BaseModel",
]
